+++ test/test_histogram.cc
@@ -5,9 +5,9 @@
 #include <cstdio>
 
 TEST(HistogramTest, Update) {
-  int N = 10;
+  unsigned int N = 10;
   Histogram hist(N);
-  for (int i = 1; i <= 2 * N; ++i) {
+  for (unsigned int i = 1; i <= 2 * N; ++i) {
     hist.Update(i, 1);
     ASSERT_EQ(std::min(i, N), hist.get_num_bins());
   }
@@ -18,7 +18,7 @@ TEST(HistogramTest, Update) {
 };
 
 TEST(HistogramTest, BenHaimAppendixA) {
-  int N = 5;
+  unsigned int N = 5;
   Histogram hist(N);
   std::vector<Histogram::Bin> bins;
   std::vector<int> sequence = {23,19,10,16,36,2,9,32,30,45};
@@ -56,7 +56,7 @@ TEST(HistogramTest, BenHaimAppendixA) {
   ASSERT_NEAR(3.275, M15.y, 0.01);
 
   std::vector<double> uniform_points = hist.Uniform(3);
-  ASSERT_EQ(2, uniform_points.size());
+  ASSERT_EQ(2, (int)uniform_points.size());
   printf("-------------\n");
   for (unsigned int i = 0; i < uniform_points.size(); ++i) {
     printf("%f ", uniform_points[i]);
